requirements.txt
Django==1.8.2
beautifulsoup4==4.4.0
dj-database-url==0.3.0
dj-static==0.0.6
django-autoslug==1.7.2
django-bower==5.0.4
django-meta==0.3.1
django-settings-export==1.0.3
gunicorn==19.3.0
kombu==3.0.26
paramiko==1.15.2
psycopg2==2.6
pycrypto==2.6.1
pystache==0.5.4
pytz==2015.2
requests==2.6.2
six==1.9.0
static==1.1.1
